Fixed auto, Add logging for commands, Add backup
Added a backup for cone align is the same as we used in stluis
Added logging for commands should be easy to remove if it breaks something which it shouldn't
Added a easy to remove way that the intake will home closer on the close side of the felid as it covers the apriltag camera which might become an issue on blue side where coming in from the non cable bump it might block april tags. Which we might use for cube align as it has been working and dresden is considering trying it matches.